[gfxconsole] Prepare for moving keyboard handling to a separate file
Plumb the keypress handler function as an argument through various
keyboard handling functions.  This is in preparation for moving that
code to a separate file that doesn't depend on gfxconsole/main.cpp.

Remove the check for g_active_vc in order to remove a dependency.
This check is just a micro-optimisation

Rename vc_input_devices_poll_thread() to reflect that it doesn't
actually poll.

Change-Id: I9b5aba449568ee0dfc2350767d700ae49cdf3c90