apps/media: make flog understand fidl connections
This CL leverages the ability to get 'related' koids for channel
handles to allow flog to comprehend what objects are connected
via a fidl connection. A fidl service can obtain a koid for the
local (service) end of a channel from a fidl InterfaceRequest or
a fidl Binding. A fidl client can obtain a koid for the remote
(service) end of a channel from a fidl InterfacePointer or a
fidl InterfaceHandle. The flog viewer now has the required
infrastructure to allow channel handlers to resolve fidl
connections reflected by identical koids in the logs, so that
parent/child and peer connections can be modeled easily.

Change-Id: I6cff92b20b5186af8d07a872996b1e063cb427a2