[ulib][musl] Move thread-specific data directly into struct pthread
This data has fixed size and is always allocated, so it simplifies
things to just embed it in the thread descriptor.  The old code was
optimizing memory use when pthread_key_create was never used in the
static linking case.  Since we don't support static linking at all,
we never get that savings anyway.

Change-Id: Ibd15ae3386f50c075e95d8e6004f9183b5d39cab