[devmgr] Allow devhost IOCTL_1H to return a data response.
For ioctls of the form "IOCTL_WRAPPER_INOUT", which set a handle
and expect a response, the devhost should transmit the returned
data. In the remoteio protocol, the amount of data transmitted is
determined by the "datalen" field of the mxrio message.

As can be seen in the return case of MXRIO_IOCTL, "msg->datalen" is
set with the value returned from the ioctl (which should also be,
in the non-error case, the size of the returned object).

This patch extends IOCTL_1H to have similar behavior. This is useful
for ioctls which both set a handle and get an object in return --
although there currently are none, this is a critical part of the
upcoming FIFO-based block protocol.

Change-Id: I1ee1d50f09bfd7ec7da37acaba1b0bdcccc02794