[ulib][musl] Get rid of builtin_tls, early thread-pointer setup
The only need for an initialized thread pointer during dynamic
linker startup was code paths that set errno.  Rework the
internal code to return mx_status_t values instead of setting
errno.  Then eliminate the builtin_tls buffer altogether and
don't initialize the thread pointer until all libraries are
loaded and TLS size is known.

If the malloc/calloc calls made during startup fail, they will
try to set errno and crash from a null thread pointer.  But that
is an unrecoverable failure in startup that can only happen from
extremely bogus ELF files or catastrophic system failure, so
diagnosing it with a message is not much better.  In the long
run, these calls will be replaced with uses of an internal
allocator that does not use errno.

One of the malloc calls on the startup path was to dynamically
allocate space for large numbers of ELF program headers.  Avoid
that by simply limiting support to a fixed maximum of 16 program
headers, which is more than enough for real-world ELF files.

Change-Id: Ice8098cc8ec14204a9e603646648858a045c18b3