[ulib][runtime] Use compiler magic to fetch the thread pointer
All aarch64 compilers have an intrinsic for the thread pointer,
so use it instead of inline asm.

Clang for x86-64 has a segment-based access feature via the
address_space attribute on pointer types, so use that instead
of inline asm.

The compiler can be smarter when using its own magic than when
using inline asm.

Change-Id: I6580070642382e430c605b79a49b4ae1febfc28f