[build][rodso] Fix .gnu.hash spelling
When linking with lld, it fails to match .gnu.hash section because
of the incorrect spelling and that section ends up in the code
segment inflating its size and causing a page fault at runtime.
Surprisingly, gold isn't affect by this, most likely because it
uses a different rule for placing .gnu.hash section.

Change-Id: I3111f8a673b51d29b9f2c8c9257773d94137b2f4