[gfxconsole] Fix "ESC D" (cursor down) sequence so that it scrolls
This sequence is supposed to scroll the console when the cursor is at
the bottom of the console (or at the bottom of the scroll region).

scroll_at_pos() only changes the lines below the cursor, which has no
effect in this case.  scroll_up() is what we want here.

Change-Id: Id4cd8289d4852d186a8b9ba0d87bfb949e98e898