[utest][core][futex] Fix bad assumptions in misaligned futex test
The old code implicitly assumed that a char[] buffer on the stack
would be aligned to sizeof(int), which there is no reason it has
to be.  It also implicitly assumed that the first futex in its
stack buffer would be initialized to zero, which of course is not
reliable.  When it's nonzero, the kernel checks that before
checking the alignment (which it probably shouldn't, but the test
shouldn't pass a pointer to uninitialized stack garbage anyway).

Change-Id: I972e53ae7f95c3d8b0700a18e486b47f43d89c02