apps/media: add fidl connections to player logs
This CL uses flog's ability to traverse fidl connections to add
the player=>source and player=>sinks connections to the logs. This
will become useful when the source and sink log their bindings in
a future CL. This CL also has the player log its binding for
completeness. At the moment, the player is at the top of the food
chain as far as flog is concerned, so nothing will reference that
binding for the time being.

Change-Id: I3e4cfa2c68ddcf2af27d436e501db92ba7adf472