[gfxconsole] Fix out-of-bounds array access in "insert/delete lines" ops
It's possible to do "insert lines" or "delete lines" control sequences
with an argument specifying a large number of lines -- larger than the
scroll area.  In that case, vc_tc_scroll() was calling
vc_device_invalidate() with 'y' and 'h' arguments beyond the end of
dev->text_buf (or before the start of dev->scrollback_buf).

Change-Id: Ie2d35611354b4186cfca8827dad0f65d69304d74