[gfxconsole] Fix bug: Cursor was displayed in margin but not removed
Before this change, if you write a string that wraps across lines,
gfxconsole displays the cursor in the right-hand margin but doesn't
remove it.  (As a side effect, the log console contains a black stripe
at the right-hand side, because the left-behind cursor is copied up
the screen by scrolling.)

Fix this by not displaying the cursor in the margin, in
vc_tc_movecursor().

Extend the tests so that test_wrapping() catches this:

 * Add a margin to the vc_device_t that the tests create.
 * Call vc_device_clear_gfx() so that the margins are cleared, so that
   we'll detect incremental update problems in the margins.

Change-Id: Ic0d718b39cf179b260973172f6d6373f30f43f6d