[ldso][crashlogger] Add MX_PROP_PROCESS_DEBUG_ADDR support
When aslr is turned on the address of the dynamic linker is no longer
fixed and thus tools like crashlogger and debuggers can no longer find
the address of ld.so's _dl_debug_addr. This patch adds
MX_PROP_PROCESS_DEBUG_ADDR as a way for tools to obtain this value.

MG-536 #done

Change-Id: I26abfbbf8719a0b226e86d87b506ca319be191a4