[devhost] introduce special "misc parent" device and protocol
The misc parent is the singleton device which is published at
/dev/misc and that all misc devices that are otherwise parentless
(not a child of some other bus, etc) are children of.

MX_PROTOCOL_MISC_PARENT is used to differentiate the parent
device from devices that implement MX_PROTOCOL_MISC (and are
aliased under /dev/class/misc/...), so we no longer have a
confusing /dev/class/misc/misc which is an alias of
/dev/class/misc.

Pure misc devices no longer need to be compiled into devhost.
If they have a single instruction bind program of this form:
 BI_MATCH_IF(EQ, BIND_PROTOCOL, MX_PROTOCOL_MISC_PARENT)
they will be correctly loaded be the primary devhost.

This also introduces DEV_FLAG_MULTI_BIND which allows multiple
children to bind to a single device, functionality needed for
the "misc parent" singleton.  Previously misc devices relied
on manually device_add()'ing themselves to the misc parent.

Change-Id: Ic9e647c7002178c4dd5d842274962d0918534ce0