[magenta] Add EPT support to the hypervisor.
We now take a VMO and treat that as the guest's physical address
space. Currently we map everything on setup, but we will look into
moving to an on-demand solution in the future.

Additionally, a new hypervisor library was introduced to manage the
guest mappings, and will be a place for most of the hypervisor code to
live.

Change-Id: I31412d47537412929fb248a69e2655376d4fbddb