[kernel][magenta] Cancelation
new syscall: mx_object_wait_cancel()

Allows you to cancel both object_wait_ and object_async_wait
syscalls. The existing pattern of closing the waited on handle
is preserved but the error status that both cases returns is
changed from

   ERR_HANDLE_CLOSED
to
   ERR_OPERATION_CANCELED

The previous status is kept but deprecated. In a followup
CL the rest of the magenta callers will be converted and
in a further CL the fuchsia callers will be converted.

Change-Id: I0fbbe664e4ef9d798d2068230d40fb6a36750d02