[gfxconsole] Speed up handling of "insert/delete lines" operations
It's possible to do "insert N lines" or "delete N lines" control
sequences where the argument N is huge.  Before this change, the
implementation in gfxconsole would take O(N) time.  We fix that so
that the argument N is first clamped to the height of the console (or
the scroll area).

Refactor the scrolling logic: Replace _scroll_up() and _scroll_down(),
which shift the display by one line at a time, with scroll_lines(),
which can shift by multiple lines (matching the tc->scroll()
callback).

Change-Id: I18646b6e769d577442720878cf4bedcf0dc6b62f