Fix usage of page snapshots in StoryProviderImpl.
A Page client that uses a PageWatcher must read only from PageSnapshots
obtained through the watcher, otherwise the state accessed by reading a
snapshot is unrelated to the state being notified about in the watcher
callback.

Apply this principle to StoryProviderImpl. Since the page snapshot is
now kept, all operations are changed to receive the page snapshot as
a parameter, rather than receiving the ledger and obtaining a page
snapshot from there.

Since the page snapshot may also change while an asynchronous operation
is in progress, all operations that access the page snapshot (i.e., all
operations) are now serialized in an operation queue. The separate operation
container is gone.

Some operations invoke other operations, which was possible in an
operation container but would deadlock in an operation queue. Such sub
operations are thus executed in a sub operation queue, which is a member
of the Operation that invokes them.

A subtlety there is that the super operation must not call its own
Done() in the result callback of the sub operation, because its Done()
call deletes the Operation and with it the OperationQueue in which
the sub operation executes, before the sub operation calls its own
Done(). This is solved by always continuing after a sub operation
with a PostTask(). This is encapsulated with the new SubResult()
method of Operation.

Change-Id: Ibb062b000fa2a555b4c217f79cadce5143e0dfcc