[mozart] Input Focus Part 1 (HitTest on Views)
This is step one for input focus and some concepts in this particular CL
are still intertwined and will be separated in subsequent CLs. For
instance we mix the notion of key event focus and touch event focus with
the Event Path and Input Focus events. Once we introduce the Request
Focus API in Part 2, we will distinguish between event path and
responder chain.

- Introducting hitable API on views, to let views participate
in hit testing.

- Basic algorithm uses the hit test on the scene graph to gather the
list of nodes and their associated views into a tree. Once the tree
of views is determined, we simultaneously query the views for whether
they are hit or not and to provide a list of subviews that
could receive events. For now we pick the first leaf to forward
input events to.

- Introducing input focus event to notify views whether they've gained
or lost focus.

- Bubbling up input event up the focus chain when not handled by a
view in the chain.

Change-Id: I2c16821918d60b638ffe340d74f873df9de0e1cf