Allow ApplicationLoader to produce null ApplicationPackage on error
Previously on error ApplicationLoader would produce an ApplicationPackage
with an invalid data vmo, which would fail FIDL validation.

* For a control flow directly to ApplicationRunner, this would produce a
  spurious error about a null package invalid response.
* For a control flow with the WIP delegating ApplicationLoader, this would
  produce a spurious fatal where the callback appeared to never have been
  called (unclear why FIDL validation failure would result in this).

Changing ApplicationPackage to optional for the ApplicationLoader callback
allows a simple valid error return which ApplicationRunner can gate on
gracefully.

An alternative solution would be to make the vmo data optional so that an
invalid vmo would pass validation. However, at this time that does not seem
warranted.

Change-Id: Id1d52a39036c2c1ab063718987cbbb4e32c9de97