apps/media: make decoder use input pts rate, not context's
The ffmpeg video decoder was attempting to get the PTS rate from
the ffmpeg 'codec context', if it was available. This might have
made sense if we were using the AVFrame PTS values that ffmpeg
(sometimes) generated. Instead, we're using the PTSs from the
input packets, so we should use the PTS rate from the input packets.
For some A/V files, these rates are the same, so the problem didn't
manifest until recently.

Change-Id: Ifa55cfd9b7baae0444cfad0bba1a1219b3b36cde