[build] Fix get-build-id to work with Clang
There are two issues:

(1) $(READELF) variable contains string with spaces so it has to be
quoted before passing to get-build-id script

(2) Unlike readelf, llvm-readobj prints the version stored in the
NT_GNU_GOLD_VERSION note. Unfortunatelly, when gold produces this note
it (accidentally?) places in three zero bytes at the end of the note
string. llvm-readobj prints those which causes grep to interpret the
output as a binary, not as a text (grep interprets any content that
contains zero byte as binary), so we need to force grep to always
interpret the output as text in this case.

Change-Id: I854806262c758cf12b7e4a85dc97ebd45e777139