[udev][usb-xhci] Ugly hack to workaround hang in xhci_reset_endpoint()
xhci_queue_transfer() calls xhci_reset_endpoint() to clear stalls
in stalled endpoints. Unfortunately it is unsafe to do this,
since xhci_reset_endpoint() is synchronous and xhci_queue_transfer()
can be called from a transaction completion callback running on the
XHCI stack's event thread.
So for now we add a one second timeout on synchronous commands and attempt
to clean up if the timeout is exceeded.

The real fix is to:
1) leave it up to the client to clear stalls, and
2) don't call command completion callbacks from the event thread

This patch should be reverted after a real solution is implemented.

Change-Id: I716cc710d61c05e9f3b761ab29be9a1bb21c3e5a