scripts/installer: builds in out dir, faster
dd(1) is a trickster, it makes you pick small buffers because it
pretends that it's blocks are disk related things. mkfile and truncate
are much faster and more appropriate on macOS and Linux. Falls back to
head(1) to read from /dev/zero otherwise.

The paths are output at the end of the script, as the user may want to
know about those.

Change-Id: Id76fd6369fc0197b364c11655bfca953c8bde52f