[gfxconsole] Fix a scrolling bug where the char under the cursor is erased
This fixes a bug where the character under the cursor gets erased when
we get a newline that scrolls up.

The bug was caused by the call to vc_device_invalidate() in
vc_tc_scroll().  This erases the cursor by redrawing the character at
the cursor's position.  However, at this point, textcon.cpp has
already updated the text buffer (to move lines upwards), so the wrong
character gets drawn -- this will be a space character when the cursor
is at the bottom of the screen.

Fix this by reordering the cursor invalidation and text copying: Do
the cursor invalidation first.  To do that, we replace the scroll()
callback (which doesn't modify the text buffer) with a copy_lines()
(which moves the lines in the text buffer as well as in the
gfx_surface).

Change-Id: Ib5d4f58ced5eccb7416aef6353fddfe498f76c82