apps/media: add the net_media_service
The new remote control scheme addresses the problem that the media
player cannot be told by a remote control point to play a new URL.
MediaPlayer has never been allowed to open URLs to a avoid making
it an 'ambient authority' that needs to check client permissions
to determine if a URL can be opened. Instead, a client specifies
the content to play in the form of a SeekingReader (essentially
an already-open URL). A remote control point has no way of doing
this.

The new approach is to create a new service (net_media_service)
that is intended to run in the user's context and so is naturally
subject to the capabilities of the user. net_media_service will
vend two types of agents: NetMediaPlayer and NetMediaPlayerNetProxy.
NetMediaPlayer is pretty much like MediaPlayer (it wraps a
MediaPlayer) except that it takes URLs instead of readers and has
explicit support for publishing itself via NetConnector.
NetMediaPlayerProxy is a proxy for a remote NetMediaPlayer.

This CL adds the net_media_service. Subsequent CLs will change the
example apps to use it.

Change-Id: Ifa997b78b1aaaf079744fc0c27e845d7584e1855