dash: sync the terminal history in real-time
This patch makes dash pull down remote entries added to the command
history on other devices in real-time.

Implementation: the controller becomes responsible for writing any new
remote entries to the control channel. The shell checks for any pending
messages on the control channel each time before displaying a new prompt.

In the future, we might want to move the control channel processing to a
dedicated thread.

This change should have no effect when the shell does not run under
shell controller (e.g. in debug shells or when moterm is run through
launcher).

Change-Id: Ief9ede32ab96def18fc90b2e496e90ec951b99b0