[ethernet] move concurrency handling into send
All the current ethernet drivers besides intel-ethernet have a mutex
inside their send implementation making them safe for concurrent calls.
This CL moves to depending on it, removing the mid-layer lock added in
CL 22609 and adding a lock to intel-ethernet.

As a side effect, it may fix MG-612 and NET-26.

Change-Id: Ibbec0ac1b3dcc58a2a818c7d33c31dcf745be661