Stop blitting depth buffer to screen; it is illegal.
Since updating to the Vulkan 1.0.42 SDK, the validation
layers no longer allow us to blit depth buffers directly.

Eventually, we'll want to render the depth buffer using
a quad and a fragment shader.  For now, don't show the
depth buffer in the debug overlay.

Also, the 1.0.42 SDK now complains about a format mismatch
between the multisample and resolve buffers.  Fixed.

Change-Id: I422196b93741d6c1f3001fbb5245b39581154bdb