[runtests] Update command line arguments
The old "toggle tests" command line argument require knowledge of which
tests are the default, which is unintuitive.

The new command line arguments can more easily run a subset of tests.

For example, to run "small + large" tests,

  "magenta$ runtests -S -m -L -p"

Will always work, regardless of the default.

Additionally, allow specification of a test name to run a single tests.
This can be useful for larger tests -- it'll allow manual users of
runtests to run targeted tests, which can be very useful for
large tests which may take a while to execute.

Change-Id: Ic3870bfec247dc91df76a5353b6bf17160462a23