Adding retry to net app loader
On the acer (and potentially other environments), the network service is
available before the network connection is established. Address this by
putting in a retry loop. The retry behavior is not yet configurable by the
client.

Also improving error reporting accordingly, though this is also not yet
surfaced to the client.

Change-Id: Id26fdeb15fc6137de9f8ff226b09d04c642dce2b