Add UDP support.
This is enough to get netcat working with the -u flag.

While here, this CL flips the read/write names of the socket
goroutines. In some ways both goroutines both read and write,
but the obvious way to name them is to match the name of the
libc call they support. They now do.

Change-Id: Ic0d5d60e0d629919191dde5ef2f521b3673ab560