apps/media: stop using deprecated ffmpeg stuff
Our new roll of ffmpeg deprecates a few things, including the 'codec'
field of the AVStream structure and the functions we used for the
actual decoding operations. This CL addresses that by using
AVStream's 'codecpar' instead of 'codec' and by using the new
decoding functions. The former requires new code to build stream
types from AVCodecParameters structures, which is a big part of this
CL. The latter motivated a refactor of FfmpegDecoderBase and its
subclasses, FfmpegAudioDecoder and FfmpegVideoDecoder.

Change-Id: I8275ab656e34001bd9000f8d30928f55044ca9f2