[make] tweak some of the module.mk code
no functional changes

Just sorts the information that goes into module-config.h so that different runs
generate the same output always. Seen when switching between make 3.81 and make 4.1.

Move the logic to generate the config file slightly down in the file to pick up any changes
appended as the file goes on.

Change-Id: I635f49dbc6cf90f00d90a7a5d0ea641944114a5b