[build] Use --sysroot for host link
HOST_CFLAGS and HOST_CPPFLAGS are insufficient for specifying
the sysroot location: they don't apply to link commands. When
building on Magenta, we also need to be able to find the system
libraries at link time, so put the --sysroot into HOST_COMPILEFLAGS
which applies to both src->obj and obj->exe builds.

Change-Id: I06b1cfffe9d989e4a8b5e79c46c17578bb43c98b