[udev][usb] Add API for resetting endpoints
When endpoints stalled, we used to automatically try to reset the endpoint
when the next transaction was queued on the stalled endpoint.
But that could result in deadlocks, since the reset endpoint operation
happened synchronously during the iotxn_queue.
Now we add an explicit API for resetting endpoints.

We also now return ERR_IO_REFUSED when attempting to queue a transaction
on a halted endpoint.

Change-Id: Ib7c73d24da07946e8ed5611905305b8f9f03c2e8