Support // paths in zsh.
With this you can use // to refer to the root of the fuchsia tree in
zsh. For example you could `cd //scripts` or run `//packages/gn/gen.py`.

This is implemented in tab completion by monkey-patching the _path_files
completion function and in the command-line editor by fixing up the
paths in ZLE's accept-line.

This removes the ~f-NAME hack I had before that was slow, ugly and I'm
pretty sure not used by anyone but me.

Change-Id: Id28a720aad47ef73fb6169a3aced3a8f2018ee16