[mozart] Get tests to run and fix bug in view associate table
- The use of a global environment with the conjunction of a local
variable to hold onto the message loop seemed to have prevented the
view manager client to stay alive across tests. Refactored the code a
bit to get tests to run.
- While running tests for registering associates, discovered a bug in
the view associate table which did not actually knew how to handle more
than one view associate, but since the tests run within the root
environment and the input manager is automatically added when the view
manager starts, when the tests registered a test associate the table
would fail on a check.

Change-Id: Iac5478ace1ffdb19f912b991c2e20fc2e381dd99