apps/media: add decoder logic for non-empty eos packets
This CL updates the decoder to handle a couple of cases that can
happen in theory but which do not happen in practice (because of the
behavior of the demux).

In practice, a packet produced by the demux is empty (lacks payload)
if and only if it is an end-of-stream packet. This is convenient for
the ffmpeg decoders, because ffmpeg uses an empty packet to signal
end-of-stream. However, the decoders should, in theory, tolerate
empty packets that aren't end-of-stream packets and end-of-stream
packets that aren't empty. This CL makes sure that's the case.

Change-Id: I18b3f488bfd75fd5a26c644a507a19b6a3118aae