[userboot] Prevent initial setup from using low address space
Make sure that the process started by userboot has roughly the low
half of its address space unused when it starts.  This makes it
possible for that process to use sanitizers that need to set up
their "shadow memory" area in the low part of the address space.
Only the mappings made by userboot are affected, so the reduction in
potential ASLR effectiveness applies only to the locations of the
initial ELF images (usually libc.so and the vDSO) and stack.

Change-Id: I95e1379082decae9107126843e89bf85d3ffcc28