[ulib][launchpad] Prevent process loading from using low address space
Make sure that a process started by launchpad via a PT_INTERP has
roughly the low half of its address space unused when it starts.
This makes it possible for that process to use sanitizers that
need to set up their "shadow memory" area in the low part of the
address space.  Only the mappings made by launchpad are affected,
so the reduction in potential ASLR effectiveness applies only to
the locations of the initial ELF images (usually libc.so and the
vDSO) and stack.

Change-Id: Iee051e8288f1833033f9f81cae408a66981866b0