Properly deduplicate Link instances by name.
When a Link is created with a name that already exists, create a new
connection to the existing instance rather than a new instance.

Because the Link now gets new connections other than through fidl
methods, those new connections must be deferred until the Link
instance is asynchronously initialized. We follow the pattern
established by StoryProviderImpl.

Change-Id: Ib2063c050c15f53f9aeb1296d57f629c4b44d18c