[mxtl][mutex] Make mxtl::Mutex work without MUSL
Under the hood, user-mode mxtl::Mutex uses C11 mtx_t.  MUSL defines an
extension (the preprocessor symbol MTX_INIT) which can be used to
statically initialize a C11 mutex.  C11, however, does not.

If the symbol is defined, allow mxtl::Mutex to use it so it may have a
constexpr constructor.  Otherwise, use the C11 standard mtx_init to
initialize the mutex.

Change-Id: I51c30e5b839e1e5079012e051ea27391fe7f599a