[common] Remove MutableByteBuffer and TransferContents()
Removed MutableBufferView::TransferContents() and added
BufferView::CopyContents(), removing the ambiguities that came with
the "maybe move" semantics. The idea is that in instances where
a move-only buffer is needed, interfaces should explicitly declare
DynamicByteBuffer as a parameter.

Also removed MutableByteBuffer which was unused, at least until we
know it's needed.

Change-Id: Ib0a0a87299355b36011a1ebf707cc49c1e904be0