[hci] Move FakeController into its own file
Moved the FakeController class out of command_channel_unittest.cc so
that it is reusable from other tests.

Also fixed a race condition that can occur in CommandChannel if
Initialize() and ShutDown() are called in quick succession on the same
thread. Fixed this by making sure that Initialize() does not return
until all of its asynchronous work is completed.

Change-Id: Ice9fc2d093f420b9aa8160d50d598c739febe3e8