Connect to Maxwell UserIntelligenceProvider.
MW-56

This changes the way Maxwell services are initialized by UserRunner and
how the services are then vended to Modules and Agents.

NOTE: There is an awkward service exchange in user_runner.cc between
UserIntelligenceProvider, AgentRunner and StoryProviderImpl. AgentRunner
needs a UserIntelligenceProvider to expose services from Maxwell through
its getIntelligenceServices() method. Initializing the Maxwell process
(through UserIntelligenceProviderFactory) requires a ComponentContext.
ComponentContext requires an AgentRunner, which creates a circular
dependency. Because of FIDL late bindings, we can get around this by
creating a new InterfaceRequest here (|intelligence_provider_request|),
making the InterfacePtr a valid proxy to be passed to AgentRunner and
StoryProviderImpl, even though it won't be bound to a real implementation
(provided by Maxwell) until later. It works, but it's not a good pattern.

Change-Id: Ic9112f926471ecf14db4382cf80838e76f01f8e8