Modesetting: Implement sending "native" DisplayPort Aux read/write requests
The "native" operations (as opposed to I2C operations) are used for
reading DisplayPort's DPCD, which is basically a register set.  This
can be used for reading info about the display, such as the number of
DisplayPort lanes it supports.

This renames the class I2cOverDpAux to DpAuxChannel to reflect its
larger functionality.

In later changes, I will implement DisplayPort link training, which
involves reading and writing the DPCD.

MA-150

Change-Id: Iea24a3ff43956f838ab7dcd4e57ddacbb63b4d05