benchmark: refactor to separate result computation from presentation
This patch lays ground for adding support for dashboard upload by
refactoring out the logic that computes self-contained results of the
benchmarking run (complete with labels, units, etc.) from the logic that
prints them out. This makes it possible to share this logic between the
existing console output and the upcoming dashboard upload.

No functional changes intended in this patch.

Change-Id: Ic94db6d3c5fadbc96ea221730d9fe7e911457d43