Do not panic on double-close.
Use the closing of dataHandle to signal internally when an endpoint
should be closed. This lets us keep the actual ep.Close in a single
place, and make sure it is called only once.

One result of this is there is a new error to handle from any action
on dataHandle, it can now return ErrHandleClosed (distinct from
ErrRemoteClosed) if dataHandle.Close is called during an action, or
ErrBadHandle, if an action is taken after dataHandle.Close returns.
All other goroutines working with the socket can treat these new
errors as a signal to stop what they are doing, so handling is
straightforward.

Change-Id: I7172df55efaa5700ce29137c3ed4733e840f5282