[devmgr] Fix MemFS rename bug
Devmgr dnodes are variably sized (due to variably sized file names).
As a consequence, when renaming from one vnode to another,
a directory dnode (carrying the name of the directory) needs to be
relocated to a new entry.

Dnodes carry a "children" list, which may or may not be empty
at the time of rename. Previously, we discarded those children.
This patch fixes this behavior to instead move them to the
newly allocated dnode.

Change-Id: I2735470297d9fe61c313a0ab387cdcfc716fa374