[kernel][panic] capsule library
With an implementation for Intel based on the RTC bank 2
non-volatile memory. Upon kernel panic we now attempt
to save the backtrace of the crash in the capsule and the
git revision of the kernel.

A capsule is memory that can survive a reboot.

This CL adds two new kernel commands:

k capsule <index>  : dumps the capsule content (index 0)
k prevpanic        : dumps the last panic saved

Notes:
- Sometimes the bios complains of a "cmos checksum"
error. It is unknown how to trigger this.
- Sometimes not all bytes make it, specally the ones
at the end.

Change-Id: Ic626bf2c919dfbc15dcc55dbdf9fd93add207693