apps/media: cancel async waits in destructors
Async waits that are left around when the fd in question is closed
cause the message loop to DCHECK (it doesn't like getting BAD_HANDLE
status from its wait). This CL makes sure that all such waits are
cancelled so this doesn't happen. Someday, we'll get a better
AsyncWaiter that'll clean up after itself (and not require static
callbacks), but this will prevent some explosions in the mean time.

Change-Id: If0ca26999bba31a59fb01d32d0b949f3d14a2651