apps/media: close bindings to prevent fidl DCHECKS
Fidl DCHECKs when an uncalled callback is destroyed and the
associated channel is still open. This can happen when an interface
implementation is destroyed when something that's holding a callback
gets destroyed before the binding does. This CL makes sure that
bindings get closed before members are destroyed so this doesn't
occur.

Change-Id: Ibdaeaf8ee6b6425af719347ed11726e477061f5c