[udev][usb-bt-hci] Remove wrong assert in ACL write code path
There used to be an assertion on the availability of iotxn's for
ACL write tasks (host --> controller). This patch fixes it so that
packets are kept in the channel buffer if there are no iotxn's
available and processed later when they do become available.

Also added code that manages the life-cycle/ownership of the ACL
channel read thread task.

Change-Id: I76f319508975788d8c52851580d2fc7544125199