Fix incremental builds of netstack2
The configuration for building netstack2 was confusing incremental
builds because godepfile was writing out $root_out_dir/netstack.d
assuming it was building a program called netstack2.  Instead of trying
to incrementally build netstack2 into a different location from its
default, this changes the build to always build netstack1 and netstack2
into their own distinct locations when desired and adds a copy rule
that copies the selected binary into $root_out_dir/netstack.  GN's
copy() rule emits ninja rules that correctly represent this copy step
so it runs when the configuration is changed or when the source binary
changes.

Change-Id: I330d1381d29b209124f0856d1328f29e62e02b9c