[ddk][iotxn] Reduce overhead of cloning iotxns
We no longer duplicate and remap the VMO handle when cloning txns.
This is safe to do as long as we guarantee that the cloned txn
is completed and its copy of the buffer is not accessed
before its parent is released.

Added asserts to ensure that cloned iotxns are only used
within these new constraints.

Also fixed a place where the usb-xhci driver was calling physmap
on iotxns that have no data

Change-Id: I406d09574db780921dca5f1ae8a2c58c3e0e6eac