apps/media: add revised_media_type to packets
This CL adds an optional field, revised_media_type, to MediaPacket.
The presence of a revised media type in a packet indicates the the
packet and subsequent packets are of the specified type. The ability
to change media types mid-stream will have a number of uses down the
road. Currently, it's needed because ffmpeg H264 decoders don't
report an accurate media type until the first frame is decoded. The
video renderer needs the updated media type to know how the frames
are laid out.

A subsequent CL will contain the video decoder and video renderer
changes required to exploit this feature for H264 rendering.g

Future uses for this field include:
1) Telling a renderer about format changes during seamless source
   transitions.
2) Supporting video streams that contain explicit type transitions.

Change-Id: Ib683050c6be171d4c7e62d0ff9417c357b267c7c