[mxtl] Remove slab allocator dependency on mxcpp
The slab allocator didn't really depend on mxcpp, it only demands that
their exists a standard bring-your-own-memory placement new operator
(and does not particularly care who implements it).

Instead of bringing in a specific protoype (magenta/new.h for
usermode, or new.h for the kernel), just manually supply the standard
prototype.

Change-Id: I595fb9933de700cdc8644414f1c8192dea60e458