Add resource files to mkbootfs.d so editing them recreates the bootfs
Package configurations list a set of binaries and resources that should
be included in the user bootfs. The binaries listed should all be built
by the GN labels defined by the package, but the resource files are
assumed to come from the source directory and not a build step. Thus,
we have to tell ninja that changes to these resource files should
result in the build step running again which we can do in the generated
depsfile for the mkbootfs step.

Fixes FW-126

Change-Id: I116b1f903b68eacbd49a573d68278f66a88bb298