[kernel][magenta] port missed signal fix
This CL introduces a semaphore as well.

The issue is that event.Signal() can be called N
times in a row to no effect except setting a flag
to true and then N threads arrive to wait on
the event but only one gets woken up.

This is not evident today because it requires
a scheduler induced delay at the right moment
and threads that eventually exit. See PS#2 for
the exact conditions.

Change-Id: Ifceec1d8c6f2c0fd9e35ab4e56d6f49bdbe031ad