[fs] Shut down filesystem if mount error occurs
This fixes an issue that may be observed while running:

  > ps
  > runtests
  > ps

And noticing that runtests actually leaves several orphaned
MinFS filesystem servers opened. After this patch, no
orphaned MinFS servers are left running.

If, for whatever reason, the parent (mount-ing) filesystem
cannot mount the child filesystem server, it should send a
signal to the child filesystem server to shut down. Previously,
the client library (fs-management) attempted to do this, but
if the library already managed to send the mount point to the
parent filesystem, then it would send an "UNMOUNT" signal to an
invalid handle.

Change-Id: Ief0d8726c363d690140b7ca3a18207941dfb2704