Don't use fidl serialization for link storage.
Just store the json value of the link directly. We may switch
to something more structured later, to accomodate metadata.

Use a key prefix for links, to allows other types of keys
in the story ledger page, and also to prevent us from trying to
read old values in existing ledgers.

Change-Id: If892dcb38b3c36c0d04c8665f9a0a6a044f045d8