[musl] Add implementation of __cxa_thread_atexit_impl
C++ ABI expects this function to be provided by the C library, libcxxabi
it has its own implementation that is used in case when C library
doesn't provide one, but that implementation has number of limitations
and pulls in additional dependencies which are undesirable
(__dso_handle).

This change adds implementation of __cxa_thread_atexit_impl into our C
library. This version is resembles the implementation __cxa_atexit, and
is significantly simpler than implementations of
__cxa_thread_atexit_impl in other C libraries because our C library
doesn't support dlclose.

MG-655

Change-Id: I1f79853f9cff3b8ecf9c35419f74bf3463f3df58