MessageQueueManager stores to a Page.
Give a Page to MessageQueueManager. The Page has a well known
name resolved by UserRunner in the user ledger. MessageQueueManager
no longer abuses the ledger repository to use a whole Ledger for
itself.

The name of the page must still be exactly 16 chars. We might switch
to a system assigned page ID stored in the root page in the future.

Collateral clanup:

* include

Change-Id: If41d9623850db8fb56396d39bc78e286e2252d0e