The Operation interface doesn't need to be public.
The Operation passes itself to the operation container,
therefore it doesn't need to be an Operation outside itself.

Also, Run() is only invoked through the Operation interface
from the operation container, so it doesn't need to be public
either.

This clarifies that the public interface of an Operation is
really its constructor only.

Change-Id: Ia0c34308a87bea3c4aee07ba989fe354e96a9697