Add a simple and experimental device map service.
The service is exposed in the user environment and gives
access to the set of device names in the root page of the
ledger.

Change the device map to be written as one key per device,
and from the device map impl, no longer from story provider
impl. With that, story provider impl no longer takes the
device name as parameter.

Implements the device map service as a complete page watcher
service. This should also give some insight how to simplify
this.

Access the device map service from example_recipe.

Change-Id: I081704dbf5167275f092c5cf99b9802786b3ebfc