[kernel][x86] Do stack_guard initialization in assembly
Any C/C++ function might wind up using the stack-guard word.
This breaks if the stack-guard word is changed during the run
of any such function.  So only change the value in assembly code
for the main CPU.  For secondary CPUs, we can install the new value
from C code only because it's in a function that will never return.

Change-Id: Ib930d0533c1d68064192d28c37154e7db0adbf4f