app: add a timeout for the initial sync
This ensures that if the initial network request hangs (neither succeeds
nor return an error), Ledger will print a warning and default to using
local data in 5s. When the request finally succeeds, it will sync as
normally.

In the future we might want to further reduce this timeout.

LE-178 #done

Change-Id: I17917b44cca297f98e41c460ff92ef19c8cd9c67