[kernel][intel_rng][debug] Clean up stack-guard initialization
The hw_rng_get_entropy code (dev/intel_rng) is no longer called
before full stack setup, so it does not need the __NO_SAFESTACK
annotations.  Move the code to choose the stack-guard value into
a C function so the part that must be in assembly is short and
simple (and this function will be reused for other machines).

Change-Id: Ib0b27bd4388883f580ecc2429e56e8235beafa1f