[build] use '.', not '-' to add "sub-module" postfixes
Since '-' is a common feature of directory names in Magenta, it's
not possible to easily figure out if ulib/foo-bar refers to a module
named foo-bar or a module named foo with a sub-module named bar.

I want to enforce that modules *must* exist in a matching directory
(as the build expects) and that sub-modules *not* have a matching
directory (so there's never a case where we could have a collision in
the build-*/... space between a module and a sub-module.

Using a character not currently used in source directory names will
allow me to forbid its use in source directory names and ensure that
there is no ambiguity.

Change-Id: Ibf3600b33c70e6920c6c52ffc1c7139f2658034a