[ddk][iotxn] replace scatter list with page array
iotxn_physmap() will perform a lookup of physical pages if
it is unset in the iotxn. It no longer returns a scatter list.

The results of the lookup are set in the new fields, 'phys',
'phys_offset', and 'phys_length'. These fields can also be
optionally set by the current owner of the iotxn.

'phys' is an array of physical pages backing the iotxn's vm
object. 'phys_offset' is the offset into the vmo the first
entry in 'phys' corresponds to, and must be aligned to
page size. 'phys_length' is the number of entries in 'phys'.

The iotxn's processor should take care of indexing into the
appropriate entry in the 'phys' array and adjusting for
unaligned offsets using the 'vmo_offset' and 'phys_offset'
fields.

Change-Id: I60a1d20bc693a3750f51bf8f92f2e5357fea147d