[tools] add `netaddr` command to find and print an ipv6 address
It uses the netboot protocol. By default it will print the magenta
address but pass --fuchsia to print the Fuchsia netstack's address.

If a hostname is supplied then the address of that host will be used
otherwise it will return the first host that responds.

This is designed so that you can use it in scripts or command-lines
like:
  ssh fuchsia@`netaddr --fuchsia`

Change-Id: I370749df9c0d20ab5fb32fbbba8f435e21617231