[ramdisk][blktest] Fix flake from small VMOs
The "unaligned request" test, existent in both the
ramdisk and blktest tests, tests the FIFO-based block
protocol when invalid alignment requests are sent.

It tests:
- Unaligned length requests (+/-1) (error expected)
- Unaligned device offset requests (error expected)
- and unaligned vmo offset requests (+1) (error NOT expected)

HOWEVER, when we create VMOs, we create them with
randomly generated sizes: 512 + (rand() % 5) * 512.

As a consequence, if we attempt to access a randomly small vmo in the
final test case (listed above), with:

- Device Offset: 0
- Vmo Offset: 1
- Length: 512
- Total VMO Size: 512

We end up trying to read beyond the end of the VMO.  This triggers the
"ERR_INVALID_ARGS" (-10) result on the block device server, which checks
that the VMO operation can act on the VMO of the requested size.
(See: udev/block/server.cpp, line 129)

As a fix, when creating a VMO for this test, specify a minimum size of
"kBlockSize * 2", so that if we read "kBlockSize" bytes from offset 1,
we always read within the bounds of the VMO.

Change-Id: Id6227c715c4e1b295a7fd111631f9b6e4f8301fa