Adding kronk_dev gn arg to select kronk-dev agent
Kronk has two version deployed to a Google Cloud bucket: a stable version for
normal use and a dev version for active development. This gn arg allows us
to switch between the two without needing a code change (which would need to
be reverted prior to each check-in).

Also updated the README accordingly.

Change-Id: I8906aff25f029987f2a2c383ea5c02385012c99b