[udev][usb-bt-hci] Close channels in hci_unbind
Moved the channel clean-up from hci_release to hci_unbind. This fixes
a leaked ref from user space when the device file is held open and
only closed when one of the HCI channels gets closed.

Also fixed a crash where thrd_join() was getting called on a NULL
thread.

Change-Id: I6926d4f251443578f547edd25bce5a2871e66b2c