[rust] Do "cargo check" for library targets
Doing a "cargo build" on a library target generates a .rlib artifact,
but that's not actually used when a library is a dependency from
another target; cargo rebuilds from source.

This change makes library targets just do "cargo check", which is a
lot faster, especially for release builds.

This is the second attempt. For library targets, it relies on the
"libfoo.rmeta" output of "cargo check", which is generated and also
listed in the depfile so the dependency can be accurately tracked.

TO-233 #done

Change-Id: I7b77d5b039912903e0b91666700e71e084f1402b